Move config helper to config package so that it wont course cycle import
