+++ config/config.go
@@ -1,4 +1,4 @@
-package helpers
+package config
 
 import (
 	"github.com/gin-gonic/gin"
